Added all functionality to the RemoveChangeController & AddChangeControl
Added the ability to be on a broadcast object in the landing page, which will then be transferred to the removechange-stage, allowing to fetch data on the cast objects, as well as choose them from the list. 
Added the searchresultchosen handlers to allow people to choose items from the listview. 
Additionally added the searchButton handler which allows for earching into the database on names of cast members, as well as the ability to choose a single object from the list which will show the data into the textfields. Chosen object can then be assigned to the given broadcast that was chosen on the landingpage. 
Minor change to the return type of unassignCast in the BroadcastClass, to allow for easier errormessaging to the user.