Added comments and todo lists and a prototype fix
Added some code in System for makeObjects that can make a hashmap with a cast as the key and the string roles as value. Needs to be talked about if we should implement it.
Added todo list in addassigncontroller, landingpagecontroller, logincontroller, system, user og cast. 
Added comments in user, system og cast.