Added a few functions. Details in description
Changed the save methods in Cast, Broadcast, Production so it takes an int value, finds out if its -1(something didnt go according to plan). If not everything went according to the plan and it will update this instance ID to match. 
Presentation layer will also update the arraylist and reset the listview so the newly added Cast, Broadcast and Prodution shows up and can be interacted with. 

Changed the interface for IPersistence to match so the methods createCastInDatabase, createBroadcastInDatabase, and createProductionInDatabase now return an int instead of a boolean. It either returns the ID or -1. 

Refactored the initilaze methods so its now 4 methods calling and passing a file to a 5th method. The 5th method (initilazeReadFromFile) returns the ID and the 4 other methods sets that ID. 
Saved us 54 lines of code

Did a little clean up in presentation so all the methods for creating and deleting also refreshes the list