This is a long one. Read description. Domain version 2 almost done
BroadCast,, movie, Production, ProductionCompany
Changed the methods for assign and unassign. They now call the update method in database while they save the state from before so they can revert back in case of a fault. 
Added update method that takes the new values, saves the old in temp variables and overwrites the current values on the instance. If it fails to update in database it reverts back to the temp variables and no changes are made. 
Added delete method that just call method in persistens and returns false or true. 

Cast>
Added update method that takes the new values, saves the old in temp variables and overwrites the current values on the instance. If it fails to update in database it reverts back to the temp variables and no changes are made. 
Added delete method that just call method in persistens and returns false or true. 
Merge method works, but there is a todo on it to figure out what to do after the merge

BroadCast, Movie
Added methods for creating the hashmap for its cast and their roles. Its done by calling a method that search for their(movie and broadcast) ID

Production
Added method for creating and getting the data for the broadcast arraylist that the production has. Calls the search method in CredITSystem to get a list over all the broadcast. 
While it does that it also figures out the highest seasonnumber by a foreach loop in the same method. 

ProductionCompany
Added methods for creating the objects in the movielist and productionlist. 

CredITSystem
Added search methods for production, productionCompany, and movie. Works getting the list from the persistents layer and then calling the private methods that makes the objects. 
Added methods(make-Movie,Production,ProductionCompany) for making the objects with the Strings that gets returned from the persistenslayer. 
Updated the make objects methods for cast, broadcast and production so it fits with the new design. 
Cleaned up the code, and used regions. 
Removed code in makeBroadCast objects from version 1 thats no longer relevant. 

IPersistence
Added some search and update methods for productioncompany, movie and production that wasnt there.