See description. Changed procedures and triggers for the database.
I strongly recommend entirely re-creating the database from the SQL file. The old procedures and triggers can cause problems if they haven't been removed fully from your database.

The actual change to the procedures is rather significant. It now checks on the ID rather than the name of the broadcast. This allows the broadcast to have any name and still be accounted for in the production table.