Added 2 forgotten methods in User and removed IOexpetions in domain
Added create productioncompany and getproductioncomapny in User class. 
Removed the IOexpection in domain on the save methods. 
Refactored the save methods to be 3 lines only. 
Changed what the methods createNewBroadcast returnerede so it returned the ID of the new broadcast, as agreed on so it could be tested if the method worked.