** Very important update ** Update your database to the newest version
If you don't update your database to the newest version in the sql file, none of the remove methods will work.

Implemented most of the remove methods, removeProductionCompany() is still missing.

Changed the 'B' in updateProductionCompanyInDataBase to lowercase.

Implemented updateCastInDatabase and updateProductionCompanyInDatabase

Added a getId() method to ProductionCompany, as it was missing.