Flair changes
Basically added some real nice to haves and not at all need to haves. Now possible to see broadcasts attached to a production in the Production-view. Can then choose a broadcast and go to the broadcast page. Also possible to see all broadcasts on a production if a production has been chosen.
Same thing goes for ProductionCompany that can show you to a Production or a Movie, depending on what you have chosen in the list.