Long description so see the description. Laust's spotted problem fix
It is now possible to navigate from a production company to a production/movie and make changes to/delete that object. To make this work, also made a check on the update for the listview to ensure no nullpointers thrown. 
Same goes for moving to a broadcast from a production.