Fixed issue with Roles not being shown for guests
The standardised method for dealing with iterating through the movie/broadcast role HashMap used the chosenMovie from the ModifyMovieController scene, and not the movie that was chosen in the guest page, therefore always using null. 

Now, to fix the issue, the method takes an IMovie/IBroadcast parameter, which allows for setting the desired chosenMovie/chosenBroadcast, therefore eliminating the issue.